Refactored UI Asset Namespaces
