+++ DEFCON Z/Assets/Scripts/Player/Player.cs
@@ -1,7 +1,7 @@
-﻿using System.Collections;
+﻿using DefconZ.UI;
+using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.UI;
 
 namespace DefconZ
 {
